% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_alerts.R
\name{alert_file_created}
\alias{alert_file_created}
\alias{alert_folder_created}
\title{Issue an alert that a file or folder has been created at the provided path.}
\usage{
alert_file_created(file, base_dir = ".", action = "")

alert_folder_created(folder, base_dir = ".", action = "")
}
\arguments{
\item{file}{(string) an absolute path to the file in question.}

\item{base_dir}{(string) a path (abs or relative) to root output directory
(defaults to ".", the project root).}

\item{action}{(string) text stating how the function reacts to the message
(default to an empty string).}

\item{folder}{an absolute path to the directory in question.}
}
\description{
\code{alert_file_created} issues an alert that a file has been created.

\code{alert_folder_created} issues an alert that a folder has been created.
}
\seealso{
Other cli-dev: 
\code{\link{alert_opening}()},
\code{\link{alert_writing_to}()},
\code{\link{backup_file}()},
\code{\link{format_action}()},
\code{\link{warn_file_exists}()},
\code{\link{warn_file_missing}()}
}
\concept{cli-dev}
