% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_alerts.R
\name{alert_writing_to}
\alias{alert_writing_to}
\alias{alert_reading_from}
\title{Notify the user of the directory being read from or written to.}
\usage{
alert_writing_to(folder)

alert_reading_from(folder)
}
\arguments{
\item{folder}{an absolute path to the directory in question.}
}
\description{
\code{alert_writing_to} gives the base output folder to which files are written.

\code{alert_reading_from} gives the base input folder from which files are read.
}
\seealso{
Other cli-dev: 
\code{\link{alert_file_created}()},
\code{\link{alert_opening}()},
\code{\link{backup_file}()},
\code{\link{format_action}()},
\code{\link{warn_file_exists}()},
\code{\link{warn_file_missing}()}
}
\concept{cli-dev}
