% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{append_skills_to_bullets}
\alias{append_skills_to_bullets}
\title{Append a link field to a title field with custom text.}
\usage{
append_skills_to_bullets(data, ix)
}
\arguments{
\item{style}{style of link, either "markdown", "latex", or "txt" (plain)}

\item{macro}{macro to use for "latex" references}
}
\description{
Append a link field to a title field with custom text.
}
\seealso{
Other prepare-utils: 
\code{\link{make_markdown_contacts}()},
\code{\link{omit_hidden_fields}()},
\code{\link{sort_skills}()}
}
\concept{prepare-utils}
\concept{prepare-utils
Append skills to a description field with matching index.}
