% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{as_filename}
\alias{as_filename}
\title{Construct a file name given a folder path, base name, id, and extension.}
\usage{
as_filename(path, basename, ext, id = NA, sep = "_")
}
\description{
Construct a file name given a folder path, base name, id, and extension.
}
\seealso{
Other build-dev: 
\code{\link{construct_app_metadata}()},
\code{\link{copy_files}()},
\code{\link{create_rprofile}()},
\code{\link{create_rproject}()},
\code{\link{get_latest_entry}()},
\code{\link{get_valid_opts}()},
\code{\link{in_tmp_env}()},
\code{\link{open_doc_and_wait}()},
\code{\link{open_with_application}()},
\code{\link{set_project_paths}()},
\code{\link{str_to_filename}()},
\code{\link{truncate_string}()},
\code{\link{validate_id}()},
\code{\link{validate_status}()},
\code{\link{write_app_metadata}()}
}
\concept{build-dev}
