% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCV-package.R
\docType{package}
\name{autoCV-package}
\alias{autoCV}
\alias{autoCV-package}
\title{autoCV}
\source{
\itemize{
\item{\href{Create Your Resume with Pagedown Package in R}{https://towardsdatascience.com/create-your-resume-with-pagedown-package-in-r-123ca6310d52}} 
\item{The rise of AutoCV: Why AutoCV will be a game changer?: https://revca-technologies.medium.com/the-rise-of-autocv-why-autocv-will-be-a-game-changer-7f7fa8dd52c0} 
\item \href{https://github.com/rstudio/pagedown/issues}{pagedown Issues}
\item \href{https://towardsdatascience.com/create-your-resume-with-pagedown-package-in-r-123ca6310d52}{Tutorial}
\item \href{https://github.com/rstudio/pagedown/issues/76}{pagedown, embedding htmlwidgets}
\item \href{https://github.com/nstrayer/datadrivencv}{datadrivencv packag} 
\item \href{https://sicss.io/2020/materials/day3-text-analysis/text-networks/rmarkdown/Text_Networks.html}{textnets vignett} 
\item \href{https://github.com/rstudio/rmarkdown/issues/2276}{font awesome version issues} 
\item \href{https://github.com/rstudio/pagedown/issues/98}{add page numbers} 
\item \href{https://github.com/poldrack/autoCV}{autoCV python pacakge} 
 }
}
\description{
Automatically generate and style your CV from tables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bschilder/autoCV}
  \item Report bugs at \url{https://github.com/bschilder/autoCV/issues}
}

}
\author{
\strong{Maintainer}: Brian Schilder \email{brian_schilder@alumni.brown.edu} (\href{https://orcid.org/0000-0001-5949-2191}{ORCID})

}
\keyword{internal}
