% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{backup_file}
\alias{backup_file}
\title{Transfer existing files to a backup directory Before overwriting.}
\usage{
backup_file(path)
}
\description{
Transfer existing files to a backup directory Before overwriting.
}
\seealso{
Other cli-dev: 
\code{\link{alert_file_created}()},
\code{\link{alert_opening}()},
\code{\link{alert_writing_to}()},
\code{\link{format_action}()},
\code{\link{warn_file_exists}()},
\code{\link{warn_file_missing}()}
}
\concept{cli-dev}
