% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCV-docs.R, R/build_.R
\name{build_}
\alias{build_}
\alias{build_title}
\alias{build_toc}
\alias{build_footer}
\alias{build_summary}
\alias{build_network}
\alias{build_skill_bars}
\alias{build_skills_plot}
\title{Build functions}
\usage{
build_title(
  name = "Leonardo da Vinci",
  logo = get_logo(),
  tagline = "Curriculum Vitae",
  extra = list("### My Affiliation", "### My Current Job", "### My Degrees"),
  logo_position = c("right", "left", "bottom")
)

build_toc(items = NULL, div = "h4", collapse = "<br>")

build_footer(
  add_github = "https://github.com/bschilder/autoCV",
  add_pagedown = FALSE,
  add_date = TRUE,
  img_width = "15px",
  sep = "<br>"
)

build_summary(
  wd = "./",
  files = get_data(dir_manual = wd, subdir = "cv_data"),
  items = c("n_years_experience_research", "n_publications", "n_preprints", "n_packages",
    "n_databases", "n_talks", "n_years_experience_teaching"),
  plus = list(n_years_experience_research = "+", n_publications = "", n_preprints = "",
    n_packages = "", n_databases = "", n_talks = "", n_years_experience_teaching = "+"),
  collapse = "<br>"
)

build_network(
  files = list.files(path = file.path("cv_data"), pattern = ".csv$", full.names = TRUE),
  min_count = 2,
  min_nchar = 2,
  layout = "layout_as_star",
  center = tolower("BM Schilder"),
  shape = "hexagon",
  randomSeed = 2023,
  save_path = NULL,
  show_plot = TRUE,
  height = 700,
  width = 1200,
  export_type = "png"
)

build_skill_bars(percent, title = "")

build_skills_plot(
  wd = "./",
  file = file.path(wd, "cv_data", "skills.csv"),
  types = NULL,
  label_alpha = 0.8,
  fill_alpha = 0.8,
  polar = FALSE,
  show_plot = TRUE,
  save_path = NULL,
  ...
)
}
\arguments{
\item{name}{Your name to put on the CV.}

\item{logo}{Logo to use.}

\item{tagline}{Tagline to use.}

\item{extra}{Extra HTML/markdown elements to append.}

\item{logo_position}{Position of the logo relative to the title.}

\item{items}{A subset of items to get from the \code{autoCV::icon_dict}.}

\item{div}{HTML div class.}

\item{collapse}{an optional character string to separate the results.  Not
    \code{\link[base]{NA_character_}}.}

\item{add_github}{Add link to \pkg{autoCV} GitHub repo.}

\item{add_pagedown}{Add \pkg{pagedown} reference.}

\item{add_date}{Add date CV last updated.}

\item{img_width}{Image width.}

\item{sep}{a character string to separate the terms.  Not
    \code{\link[base]{NA_character_}}.}

\item{wd}{Working directory where the "cv_data" and "img" subfolders
are stored.}

\item{files}{Named list containing paths to CV data files.}

\item{plus}{Extra symbols (e.g. "+") to include at the end of
each respective element.}

\item{min_count}{Minimum word count.}

\item{min_nchar}{Minimum word length.}

\item{layout}{: Character Name of igraph layout function to use. Default to "layout_nicely"}

\item{center}{Center node ID.}

\item{shape}{: String. Default to 'ellipse'. The shape defines what the node looks like. There are two types of nodes. One type has the label inside of it and the other type has the label underneath it. The types with the label inside of it are: ellipse, circle, database, box, text. The ones with the label outside of it are: image, circularImage, diamond, dot, star, triangle, triangleDown, hexagon, square and icon.}

\item{randomSeed}{: Number. The nodes are randomly positioned initially. This means that the settled result is different every time. If you provide a random seed manually, the layout will be the same every time.}

\item{save_path}{Path to save to.}

\item{show_plot}{logical. If TRUE, the plot will be shown.}

\item{height}{: String. Default to "100\%". The height of the network in pixels or as a percentage.}

\item{width}{: String. Default to "100\%". The width of the network in pixels or as a percentage.}

\item{export_type}{Type of export. One of "png" (default), "jpeg" or "pdf".}

\item{percent}{Percent skill value.}

\item{title}{Title.}

\item{file}{Data file path.}

\item{types}{Subset the `Type` column.}

\item{label_alpha}{numeric. The alpha value for the labels.}

\item{fill_alpha}{numeric. The alpha value for the fill.}

\item{polar}{logical. If TRUE, the plot will be polar.}

\item{...}{one or more \R objects, to be converted to character vectors.}
}
\value{
Formatted data.
}
\description{
Functions to build complex elements from CV data.
}
\section{Functions}{
\itemize{
\item \code{build_title()}: build_

\item \code{build_toc()}: build_

\item \code{build_footer()}: build_

\item \code{build_summary()}: build_

\item \code{build_network()}: build_

\item \code{build_skill_bars()}: build_

\item \code{build_skills_plot()}: build_

}}
\concept{build_}
