% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_base_directory}
\alias{build_base_directory}
\title{Create the file tree and necessary data files to begin building a resume.}
\usage{
build_base_directory()
}
\description{
Create the file tree and necessary data files to begin building a resume.
}
\examples{
in_tmp_env({
  build_base_directory()
})

}
\seealso{
Other cli: 
\code{\link{build_app_directory}()},
\code{\link{get_app_info}()},
\code{\link{open_app}()},
\code{\link{render_app}()},
\code{\link{render_cover}()},
\code{\link{render_cv_as_html}()},
\code{\link{render_resume}()},
\code{\link{update_app_info}()}
}
\concept{cli}
