% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_cv.R
\name{build_skill_bars}
\alias{build_skill_bars}
\title{Construct a bar chart of skills.}
\usage{
build_skill_bars(skill_data)
}
\description{
Construct a bar chart of skills.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(ggplot2)
data("example_skill_data", package = "autocv")
technical_skills <- filter(example_skill_data, key_competency == "x")
build_skill_bars(technical_skills)

}
\references{
Strayer N (2020). datadrivencv. R package version 0.1.0.
}
\seealso{
Other build: 
\code{\link{build_skill_list}()},
\code{\link{build_skill_table}()}
}
\concept{build}
