% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{build_skill_list}
\alias{build_skill_list}
\title{Build skill list.}
\usage{
build_skill_list(
  skill_data,
  sep = c(" • ", "\\n"),
  bullet_style = c("", "-", "+"),
  bold_headers = TRUE,
  separate_competencies = TRUE,
  competencies_header = "Other"
)
}
\description{
Build skill list.
}
\seealso{
Other build: 
\code{\link{build_skill_bars}()},
\code{\link{build_skill_table}()}
}
\concept{build}
