% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_resume.R
\name{build_skill_table}
\alias{build_skill_table}
\title{Build a two-column table of skills.}
\usage{
build_skill_table(skill_data)
}
\arguments{
\item{skill_data}{A data frame containing skill data.}
}
\value{
A LaTeX string.
}
\description{
Returns a string of text that can be rendered as LaTeX.
}
\examples{
# Define some data ----------------------------------------------------------
library(dplyr, warn.conflicts = FALSE)
skill <- c("R", "SQL", "Excel", "Pandas", "Feature Engineering")
alias <- c("Coding", "Coding", "Data Analysis", "Data Analysis", "ML")
skill_data <- dplyr::bind_cols(alias, skill)
colnames(skill_data) <- c("alias", "skill")
print(skill_data)


# Render --------------------------------------------------------------------
skill_table <- build_skill_table(skill_data)
paste(skill_table)

# With a larger dataset -----------------------------------------------------
data("example_skill_data", package = "autocv")
skill_table <- build_skill_table(example_skill_data)
paste(skill_table)

}
\seealso{
Other build: 
\code{\link{build_skill_bars}()},
\code{\link{build_skill_list}()}
}
\concept{build}
