% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{cli_apply}
\alias{cli_apply}
\title{CLI handler for 'update'.}
\usage{
cli_apply(args)
}
\description{
CLI handler for 'update'.
}
\seealso{
Other cli-handlers: 
\code{\link{cli_build}()},
\code{\link{cli_check}()},
\code{\link{cli_edit}()},
\code{\link{cli_open}()},
\code{\link{cli_render}()},
\code{\link{cli_review}()}
}
\concept{cli-handlers}
