% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{cli_check}
\alias{cli_check}
\title{CLI handler for 'check'.}
\usage{
cli_check(args)
}
\description{
CLI handler for 'check'.
}
\seealso{
Other cli-handlers: 
\code{\link{cli_apply}()},
\code{\link{cli_build}()},
\code{\link{cli_edit}()},
\code{\link{cli_open}()},
\code{\link{cli_render}()},
\code{\link{cli_review}()}
}
\concept{cli-handlers}
