% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{cli_edit}
\alias{cli_edit}
\title{CLI handler for 'edit'.}
\usage{
cli_edit(args)
}
\description{
CLI handler for 'edit'.
}
\seealso{
Other cli-handlers: 
\code{\link{cli_build}()},
\code{\link{cli_render}()}
}
\concept{cli-handlers}
