% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{cli_open}
\alias{cli_open}
\title{CLI handler for 'open'.}
\usage{
cli_open(args)
}
\description{
CLI handler for 'open'.
}
\seealso{
Other cli-handlers: 
\code{\link{cli_apply}()},
\code{\link{cli_build}()},
\code{\link{cli_check}()},
\code{\link{cli_edit}()},
\code{\link{cli_render}()},
\code{\link{cli_review}()}
}
\concept{cli-handlers}
