% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{cli_render}
\alias{cli_render}
\title{CLI handler for 'render'.}
\usage{
cli_render(args)
}
\description{
CLI handler for 'render'.
}
\seealso{
Other cli-handlers: 
\code{\link{cli_build}()},
\code{\link{cli_edit}()}
}
\concept{cli-handlers}
