% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{cli_review}
\alias{cli_review}
\title{CLI handler for 'review'.}
\usage{
cli_review(args)
}
\description{
CLI handler for 'review'.
}
\seealso{
Other cli-handlers: 
\code{\link{cli_apply}()},
\code{\link{cli_build}()},
\code{\link{cli_check}()},
\code{\link{cli_edit}()},
\code{\link{cli_open}()},
\code{\link{cli_render}()}
}
\concept{cli-handlers}
