% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{construct_app_filepath}
\alias{construct_app_filepath}
\title{Take an application config object and produce the target dir.}
\usage{
construct_app_filepath(app_info = load_job_info())
}
\description{
Take an application config object and produce the target dir.
}
\seealso{
Other cli-utils: 
\code{\link{construct_app_metadata}()},
\code{\link{load_job_info}()},
\code{\link{load_log}()},
\code{\link{str_to_filename}()},
\code{\link{validate_id}()},
\code{\link{write_app_metadata}()},
\code{\link{write_log_entry}()}
}
\concept{cli-utils}
