% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{copy_files}
\alias{copy_files}
\alias{create_folders}
\title{Copy files and create folders from lists of source and target paths.}
\usage{
copy_files(source_paths, target_paths, base_dir = ".")

create_folders(target_paths, base_dir = ".")
}
\arguments{
\item{source_paths}{a char vector of corresponding absolute source paths.}

\item{target_paths}{a char vector of target paths.}

\item{base_dir}{the root folder from which to report command line messages.}
}
\description{
\code{copy_files} copies files from source to corresponding target paths.

\code{create_folders} creates folders at the given target paths.
}
\seealso{
Other build-dev: 
\code{\link{as_filename}()},
\code{\link{construct_app_metadata}()},
\code{\link{create_rprofile}()},
\code{\link{create_rproject}()},
\code{\link{get_latest_entry}()},
\code{\link{get_valid_opts}()},
\code{\link{in_tmp_env}()},
\code{\link{open_doc_and_wait}()},
\code{\link{set_project_paths}()},
\code{\link{str_to_filename}()},
\code{\link{truncate_string}()},
\code{\link{validate_id}()},
\code{\link{validate_status}()},
\code{\link{write_app_metadata}()}
}
\concept{build-dev}
