% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_skills_report.R
\name{count_terms}
\alias{count_terms}
\title{Count keyword occurrences in a file based on a supplied list.}
\usage{
count_terms(terms, counts = NA, doc, order = c("bysource", "bycount"))
}
\description{
Count keyword occurrences in a file based on a supplied list.
}
\examples{
# Get resume counts ---------------------------------------------------------

# Load example data
data("example_posting", package = "autocv")
terms <- readLines(autocv_resources("resources/skill_list.txt"))

posting_counts <- count_terms(
  terms = terms,
  doc = example_posting, 
  order = "bysource"
)

posting_counts

}
\seealso{
Other report: 
\code{\link{run_skill_count}()},
\code{\link{run_skill_report}()}
}
\concept{report}
