% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_webpage_txt}
\alias{download_webpage_txt}
\alias{download_webpage_pdf}
\title{Save the text from a webpage url.}
\usage{
download_webpage_txt(
  url,
  output_filepath = NA,
  output_dir = "input",
  output_filename = "posting.txt",
  base_dir = "."
)

download_webpage_pdf(
  url,
  output_filepath = NA,
  output_dir = "input",
  output_filename = "posting.pdf"
)
}
\arguments{
\item{url}{The web address of the job posting.}

\item{output_file}{The desired output filepath.}
}
\description{
\code{download_webpage_txt} saves output to a .txt file.

\code{download_webpage_pdf} saves output to a .pdf file.
}
\examples{
library(rvest, warn.conflicts = FALSE)
url <- "https://en.wikipedia.org/wiki/R_(programming_language)"
fil <- tempfile("file", tempdir(), fileext = ".txt")
download_webpage_txt(url, fil)
readLines(fil, n = 1)

unlink(fil)

url <- "https://stat.ethz.ch/R-manual/R-devel/library/utils/html/download.file.html"
fil <- tempfile("file", tempdir(), fileext = ".pdf")
download_webpage_pdf(url, fil)
unlink(fil)

}
\seealso{
Other report-dev: 
\code{\link{find_matches_and_positions}()},
\code{\link{prep_term_list}()}
}
\concept{report-dev}
