% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ex_data}
\alias{ex_data}
\title{Example data}
\usage{
ex_data(dir_only = TRUE)
}
\arguments{
\item{dir_only}{Only return the directory.}
}
\value{
Paths to files.
}
\description{
Return paths to example CV data.
}
\examples{
ex_data() 
}
