% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_contact_data}
\alias{example_contact_data}
\title{Sample contact data}
\format{
A data frame with 9 rows and 5 columns:
\describe{
\item{loc}{Name of the contact info field (character).}

\item{order}{Ordinal position in which to display the item
in your resume (positive integer).}

\item{icon}{Name of a Font Awesome icon to display next to the item;
optional (character).}

\item{address_text}{Text to display for the item,
in place of the \code{address}, if provided (see below) (character).}

\item{address}{Link or path to the website, web address, or file to
redirect to when the corresponding \code{address_text} is clicked;
optional (character).}
}
}
\source{
Generated for demonstration purposes.
}
\usage{
example_contact_data
}
\description{
A synthetic dataset containing personal contact information and
associated links, including a home address, email, and personal websites.
}
\examples{
data(example_contact_data)
head(example_contact_data)

}
\keyword{internal}
