% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_job_metadata}
\alias{example_job_metadata}
\title{Sample job metadata}
\format{
A named list with x fields:
\describe{
\item{field}{description}
}
}
\source{
Generated for demonstration purposes.
}
\usage{
example_job_metadata
}
\description{
A synthetic dataset containing information about ...
}
\examples{
data(example_job_metadata)
head(example_job_metadata)

}
\keyword{internal}
