% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_position_data}
\alias{example_position_data}
\title{Sample position data}
\format{
A data frame with x rows and y columns:
\describe{
\item{field}{description}
\item{skill_1, skill_2, ..., skill_5}{description (all character)}
\item{skill_ix_iy}{ ,
where \code{ix} and \code{iy} can each be an integer from 1 to 5 (all character)}
}
}
\source{
Generated for demonstration purposes.
}
\usage{
example_position_data
}
\description{
A synthetic dataset containing information about ...
}
\examples{
data(example_position_data)
head(example_position_data)

}
\keyword{internal}
