% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_posting}
\alias{example_posting}
\title{Sample job posting}
\format{
A character vector of length 1
}
\source{
Generated for demonstration purposes.
}
\usage{
example_posting
}
\description{
A hypothetical job posting.
}
\examples{
data(example_posting)
head(example_posting)

}
\keyword{internal}
