% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_skill_data}
\alias{example_skill_data}
\title{Example skill dataset}
\format{
\subsection{\code{example_skill_data}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
example_skill_data
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{internal}
