% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{find_matches_and_positions}
\alias{find_matches_and_positions}
\title{Find the number and position of terms matching a given term in a given text.}
\usage{
find_matches_and_positions(term, text, ignore.case = TRUE)
}
\arguments{
\item{term}{A string representing the term to be matched.}

\item{text}{A string representing the text to check for matches.}
}
\value{
A named list with the following elements:
\describe{
\item{count}{Integer, the number of matches.}
\item{first_ix}{Integer, the position of the first match.}
}
}
\description{
Find the number and position of terms matching a given term in a given text.
}
\examples{
text <- "This is a sample text with several sample words."
term <- "sample"
find_matches_and_positions(term = term, text = text)

}
\seealso{
Other report-dev: 
\code{\link{download_webpage_txt}()},
\code{\link{prep_term_list}()}
}
\concept{report-dev}
