% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_alerts.R
\name{format_action}
\alias{format_action}
\title{Format any 'action' text provided to a cli alert before printing.}
\usage{
format_action(action)
}
\description{
Format any 'action' text provided to a cli alert before printing.
}
\seealso{
Other cli-dev: 
\code{\link{alert_file_created}()},
\code{\link{alert_opening}()},
\code{\link{alert_writing_to}()},
\code{\link{warn_file_exists}()},
\code{\link{warn_file_missing}()}
}
\concept{cli-dev}
