% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCV-templates.R, R/get_.R
\name{get_}
\alias{get_}
\alias{get_template}
\alias{get_data}
\alias{get_css}
\alias{get_logo}
\title{Get functions}
\usage{
get_template(save_dir = here::here())

get_data(dir_only = TRUE, dir_manual = NULL)

get_css()

get_logo(
  width = "40px",
  id = "header_img",
  text_before = NULL,
  text_after = NULL,
  link = "https://github.com/bschilder/autoCV",
  img = system.file("hex", "hex.png", package = "autoCV"),
  style = "border-radius:5\%;"
)
}
\arguments{
\item{save_dir}{Directory to save template file to.}

\item{dir_only}{Only return the directory.}

\item{dir_manual}{Specify the directory to find the files in manually.}

\item{width}{Image width.}

\item{id}{HTML element id.}

\item{text_before}{HTML element text before image.}

\item{text_after}{HTML element text after image.}

\item{link}{URL to use.}

\item{img}{Path or URL to image.}

\item{style}{HTML element style.}
}
\value{
Resources.
}
\description{
Functions to get resources and example data built into \pkg{autoCV}.
}
\section{Functions}{
\itemize{
\item \code{get_template()}: get_

\item \code{get_data()}: get_

\item \code{get_css()}: get_

\item \code{get_logo()}: get_

}}
\concept{get_}
