% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{get_app_info}
\alias{get_app_info}
\alias{get_status_report}
\alias{get_app_path_to}
\title{Access application metadata.}
\usage{
get_app_info(
  id = c("all", "latest"),
  field = "all",
  status = c("all", "ipr", "applied", "interviewed", "rejected",
    "interviewed_then_rejected"),
  app_period = "latest",
  app_dir = "applications"
)

get_status_report(app_period = "latest")

get_app_path_to(id = "latest", doc = "resume", app_period = "latest")
}
\description{
\code{get_app_info} retrieves requested info for a given application from its log.

\code{get_status_report} returns a table of proportions for each status.

\code{get_app_path_to} returns an abs path to a doc from a given application.
}
\seealso{
Other cli: 
\code{\link{build_app_directory}()},
\code{\link{build_base_directory}()},
\code{\link{open_app}()},
\code{\link{render_app}()},
\code{\link{render_cover}()},
\code{\link{render_cv_as_html}()},
\code{\link{render_resume}()},
\code{\link{update_app_info}()}
}
\concept{cli}
