% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{get_path_to}
\alias{get_path_to}
\title{Construct an absolute path to a project directory for saving and loading.}
\usage{
get_path_to(
  dir = c("src", "input", "output", "applications", "notebooks", "templates", "extdata",
    "css", "latex", "resources")
)
}
\arguments{
\item{dir}{(string) the directory name.}
}
\description{
This function constructs an absolute path to a specified project directory.
Directory paths, relative to the project root, are defined in the
user's .Rprofile environment variables.
}
\examples{
# get_path_to("src")

}
\seealso{
Other data: 
\code{\link{autocv_resources}()}
}
\concept{data}
