% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{get_valid_opts}
\alias{get_valid_opts}
\title{Obtain a set of valid options for indexing a job log by a given argument.}
\usage{
get_valid_opts(log = load_log(), arg = c("id", "field", "doc"))
}
\description{
Obtain a set of valid options for indexing a job log by a given argument.
}
\seealso{
Other build-dev: 
\code{\link{as_filename}()},
\code{\link{construct_app_metadata}()},
\code{\link{copy_files}()},
\code{\link{create_rprofile}()},
\code{\link{create_rproject}()},
\code{\link{get_latest_entry}()},
\code{\link{in_tmp_env}()},
\code{\link{open_doc_and_wait}()},
\code{\link{open_with_application}()},
\code{\link{set_project_paths}()},
\code{\link{str_to_filename}()},
\code{\link{truncate_string}()},
\code{\link{validate_id}()},
\code{\link{validate_status}()},
\code{\link{write_app_metadata}()}
}
\concept{build-dev}
