% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCV-docs.R, R/icon_.R
\name{icon_}
\alias{icon_}
\alias{icon_link}
\alias{icon_dict}
\alias{icon_sub}
\title{Icon functions}
\usage{
icon_link(link, text_before = NULL, text_after = link, icon)

icon_dict(
  items = NULL,
  dict = list(education = "graduation-cap", skills = "check", expertise = "check",
    publications = "file-text", preprints = "file-edit", acknowledgements = "file",
    reviewerships = "file", internal_talks = "person-chalkboard", invited_talks =
    "person-chalkboard", conference_talks = "person-chalkboard", posters =
    "person-chalkboard", experience = "suitcase", teaching = "chalkboard-teacher",
    packages = "wrench", websites = "computer", databases = "database", grants =
    "dollar", awards = "award", affiliations = "building-columns", 
    
    data_visualisation = "circle-nodes", extracurricular = "icons"),
  as_fa = FALSE,
  as_icon = FALSE,
  as_toc = FALSE,
  collapse = "<br>",
  icn_width = "12px"
)

icon_sub(text)
}
\arguments{
\item{link}{URL to use.}

\item{text_before}{HTML element text before image.}

\item{text_after}{HTML element text after image.}

\item{icon}{Icon name.}

\item{items}{A subset of items to get from the \code{autoCV::icon_dict}.}

\item{dict}{Icon dictionary.}

\item{as_fa}{Return as \link[fontawesome]{fa} icon id.}

\item{as_icon}{Return as HTML icon object (made with \link[fontawesome]{fa}).}

\item{as_toc}{Return as Table of Contents.}

\item{collapse}{an optional character string to separate the results.  Not
    \code{\link[base]{NA_character_}}.}

\item{icn_width}{Icon width.}

\item{text}{Text to run substitution on.}
}
\value{
Formatted data.
}
\description{
Functions to get icons for respective CV data categories.
}
\section{Functions}{
\itemize{
\item \code{icon_link()}: icon_

\item \code{icon_dict()}: icon_

\item \code{icon_sub()}: icon_

}}
\concept{icon_}
