% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCV-docs.R, R/img_.R
\name{img_}
\alias{img_}
\alias{img_get}
\alias{img_link}
\title{Image functions}
\usage{
img_get(
  name,
  dir = file.path("img"),
  width = NULL,
  height = NULL,
  as_html = TRUE,
  style = NULL,
  collapse = " ",
  default = "",
  verbose = FALSE
)

img_link(
  link,
  img,
  alt = NULL,
  width,
  class = "image",
  id = NULL,
  text_before = NULL,
  text_after = NULL,
  style = NULL
)
}
\arguments{
\item{name}{Image name.}

\item{dir}{Directory to look for image files in.}

\item{width}{Image width.}

\item{height}{Image height.}

\item{as_html}{Wrap in HTML element.}

\item{style}{HTML element style.}

\item{collapse}{an optional character string to separate the results.  Not
    \code{\link[base]{NA_character_}}.}

\item{default}{Default value when image file not found.}

\item{verbose}{Print messages.}

\item{link}{URL to use.}

\item{img}{Path or URL to image.}

\item{alt}{HTML element alt description.}

\item{class}{HTML element class.}

\item{id}{HTML element id.}

\item{text_before}{HTML element text before image.}

\item{text_after}{HTML element text after image.}
}
\value{
Formatted data.
}
\description{
Functions to create image elements.
}
\section{Functions}{
\itemize{
\item \code{img_get()}: img_

\item \code{img_link()}: img_

}}
\concept{img_}
