% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{load_job_info}
\alias{load_job_info}
\title{Read user-defined metadata for the present job.}
\usage{
load_job_info(
  field = "all",
  filename = "job_metadata.yml",
  dir = "input",
  open = FALSE
)
}
\value{
A nested named list of strings with user-specified job details.
}
\description{
Read user-defined metadata for the present job.
}
\examples{
load_job_info(filename = "template_job_metadata.yml", dir = "templates")

}
\seealso{
Other data: 
\code{\link{autocv_resources}()},
\code{\link{load_application_data}()},
\code{\link{load_log}()}
}
\concept{data}
