% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{load_log}
\alias{load_log}
\title{Load the log for a given application period.}
\usage{
load_log(app_period = "latest", app_dir = "applications")
}
\description{
Load the log for a given application period.
}
\examples{
#

}
\seealso{
Other data: 
\code{\link{autocv_resources}()},
\code{\link{load_application_data}()},
\code{\link{load_job_info}()}
}
\concept{data}
