% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{make_markdown_contacts}
\alias{make_markdown_contacts}
\alias{make_latex_contacts}
\alias{make_txt_contacts}
\title{Prepare individual entry in contact info card.}
\usage{
make_markdown_contacts(contact_data)

make_latex_contacts(contact_data, macro)

make_txt_contacts(contact_data)
}
\arguments{
\item{contact_data}{A data frame containing the contact data.}

\item{macro}{The desired LaTeX macro to use for rendering hyperlinks.}
}
\description{
\code{make_markdown_contacts} prepares an individual markdown entry.

\code{make_latex_contacts} prepares an individual LaTeX entry.

\code{make_txt_contacts} prepares an individual plain text entry.
}
\seealso{
Other prepare-utils: 
\code{\link{append_skills_to_bullets}()},
\code{\link{omit_hidden_fields}()},
\code{\link{sort_skills}()}
}
\concept{prepare-utils}
