% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_skills_report.R
\name{match_pattern}
\alias{match_pattern}
\title{Check for term matches embedded in text.}
\usage{
match_pattern(term, text)
}
\arguments{
\item{term}{Text to check.}
}
\description{
Check for term matches embedded in text.
}
\seealso{
Other report-utils: 
\code{\link{download_webpage_txt}()}
}
\concept{report-utils}
