% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCV-docs.R, R/n_.R
\name{n_}
\alias{n_}
\alias{n_years_experience}
\alias{n_tools}
\alias{n_packages}
\alias{n_webapps}
\alias{n_websites}
\alias{n_web}
\alias{n_databases}
\alias{n_rpackages}
\alias{n_publications}
\alias{n_posters}
\alias{n_talks}
\alias{n_grants}
\title{Count functions}
\usage{
n_years_experience(
  file = file.path("cv_data", "experience.csv"),
  types = "research"
)

n_tools(file = file.path("cv_data", "tools.csv"), types = NULL)

n_packages(
  file = file.path("cv_data", "tools.csv"),
  types = c("package", "web app")
)

n_webapps(file = file.path("cv_data", "tools.csv"), types = "web app")

n_websites(file = file.path("cv_data", "tools.csv"), types = "website")

n_web()

n_databases(file = file.path("cv_data", "tools.csv"), types = "database")

n_rpackages(
  file = file.path("cv_data", "tools.csv"),
  types = "package",
  languages = c("^R$", "^R,*")
)

n_publications(
  file = file.path("cv_data", "publications.csv"),
  types = "publication"
)

n_posters(file = file.path("cv_data", "publications.csv"), types = "poster")

n_talks(file = file.path("cv_data", "talks.csv"), types = NULL)

n_grants(
  file = file.path("cv_data", "grants.csv"),
  types = "grant",
  roles = c("Primary applicant", "Co-applicant")
)
}
\arguments{
\item{file}{Data file path.}

\item{types}{Subset the `Type` column.}

\item{languages}{Subset the `Language` column.}

\item{roles}{Subset the `Role` column.}
}
\value{
Formatted data.
}
\description{
Functions to compute counts from various aspects of your CV data.
}
\section{Functions}{
\itemize{
\item \code{n_years_experience()}: n_

\item \code{n_tools()}: n_

\item \code{n_packages()}: n_

\item \code{n_webapps()}: n_

\item \code{n_websites()}: n_

\item \code{n_web()}: n_

\item \code{n_databases()}: n_

\item \code{n_rpackages()}: n_

\item \code{n_publications()}: n_

\item \code{n_posters()}: n_

\item \code{n_talks()}: n_

\item \code{n_grants()}: n_

}}
\concept{n_}
