% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{open_app}
\alias{open_app}
\alias{edit_app}
\alias{edit_base}
\alias{delete_app}
\title{Open, edit, and remove application documents.}
\usage{
open_app(doc = "resume", id = "latest", app_period = "latest")

edit_app(
  id = "latest",
  app_period = "latest",
  docs = c("resume", "all", "cover", "qa")
)

edit_base(docs = c("resume", "all", "cover", "qa"))

delete_app(app_id, log = load_log())
}
\description{
\code{open_app} opens documents (resume, by default) for a given application.

\code{edit_app} opens spreadsheet data files for a given application for editing.

\code{edit_base} opens base spreadsheet data files for editing.

\code{delete_app} deletes the application folder and log entry for a given id.
}
\seealso{
Other cli: 
\code{\link{build_app_directory}()},
\code{\link{build_base_directory}()},
\code{\link{get_app_info}()},
\code{\link{render_app}()},
\code{\link{render_cover}()},
\code{\link{render_cv_as_html}()},
\code{\link{render_resume}()},
\code{\link{update_app_info}()}
}
\concept{cli}
