% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{open_doc_and_wait}
\alias{open_doc_and_wait}
\title{Open an application document and pause execution until a key is entered.}
\usage{
open_doc_and_wait(path, base_dir = ".", prompt = "Enter y to continue: ")
}
\description{
Open an application document and pause execution until a key is entered.
}
\seealso{
Other build-dev: 
\code{\link{as_filename}()},
\code{\link{construct_app_metadata}()},
\code{\link{copy_files}()},
\code{\link{create_rprofile}()},
\code{\link{create_rproject}()},
\code{\link{get_latest_entry}()},
\code{\link{get_valid_opts}()},
\code{\link{in_tmp_env}()},
\code{\link{open_with_application}()},
\code{\link{set_project_paths}()},
\code{\link{str_to_filename}()},
\code{\link{truncate_string}()},
\code{\link{validate_id}()},
\code{\link{validate_status}()},
\code{\link{write_app_metadata}()}
}
\concept{build-dev}
