% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{open_with_application}
\alias{open_with_application}
\title{Open the document at a path with a given application.}
\usage{
open_with_application(path, app = NULL)
}
\arguments{
\item{path}{Path to the file to open.}

\item{app}{Optional. Name of the application to open with (macOS only).}
}
\description{
Checks if the user is on macOS and \code{app} is supplied, and uses the 'open'
system command if so. Falls back to \code{fs::file_show()} otherwise.
}
\seealso{
Other build-dev: 
\code{\link{as_filename}()},
\code{\link{construct_app_metadata}()},
\code{\link{copy_files}()},
\code{\link{create_rprofile}()},
\code{\link{create_rproject}()},
\code{\link{get_latest_entry}()},
\code{\link{get_valid_opts}()},
\code{\link{in_tmp_env}()},
\code{\link{open_doc_and_wait}()},
\code{\link{set_project_paths}()},
\code{\link{str_to_filename}()},
\code{\link{truncate_string}()},
\code{\link{validate_id}()},
\code{\link{validate_status}()},
\code{\link{write_app_metadata}()}
}
\concept{build-dev}
