% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_bio}
\alias{prepare_bio}
\title{Prepare bio.}
\usage{
prepare_bio(text_data, use_abridged = FALSE)
}
\arguments{
\item{text_data}{A spreadsheet containing resume text data.}
}
\description{
Prepare bio.
}
\seealso{
Other prepare: 
\code{\link{prepare_description_bullets}()},
\code{\link{prepare_links}()},
\code{\link{prepare_timeline}()}
}
\concept{prepare}
