% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{prepare_description_bullets}
\alias{prepare_description_bullets}
\title{Prepare entry descriptions.}
\usage{
prepare_description_bullets(data, bullet_style = c("-", "+"))
}
\description{
Prepare entry descriptions.
}
\seealso{
Other prepare: 
\code{\link{prepare_bio}()},
\code{\link{prepare_links}()},
\code{\link{prepare_timeline}()}
}
\concept{prepare}
