% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_timeline}
\alias{prepare_timeline}
\title{Prepare dates.}
\usage{
prepare_timeline(data, order = c("chronological", "reversed"))
}
\arguments{
\item{data}{A spreadsheet containing position data.}
}
\description{
Prepare dates.
}
\seealso{
Other prepare: 
\code{\link{prepare_bio}()},
\code{\link{prepare_description_bullets}()},
\code{\link{prepare_links}()}
}
\concept{prepare}
