% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{preprocess_contacts}
\alias{preprocess_contacts}
\title{Prepare contact info card.}
\usage{
preprocess_contacts(
  contact_data,
  style = c("markdown", "latex", "txt"),
  macro = c("myhref", "href", NA)
)
}
\description{
Prepare contact info card.
}
\seealso{
Other pipeline: 
\code{\link{preprocess_entries}()},
\code{\link{preprocess_text}()}
}
\concept{pipeline}
