% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{preprocess_entries}
\alias{preprocess_entries}
\title{Run each resume data pre-processing step in sequence.}
\usage{
preprocess_entries(
  entry_data,
  style = c("markdown", "latex", "txt"),
  order = c("chronological", "reversed"),
  bullet_style = c("-", "+")
)
}
\description{
Run each resume data pre-processing step in sequence.
}
\seealso{
Other pipeline: 
\code{\link{preprocess_contacts}()},
\code{\link{preprocess_text}()}
}
\concept{pipeline}
