% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_text}
\alias{preprocess_text}
\title{Prepare text blocks.}
\usage{
preprocess_text(text_data, style = c("markdown", "latex", "txt"))
}
\description{
Prepare text blocks.
}
\seealso{
Other pipeline: 
\code{\link{preprocess_contacts}()},
\code{\link{preprocess_entries}()}
}
\concept{pipeline}
