% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover.R
\name{print_achievements}
\alias{print_achievements}
\title{Print the achievements listed in a text dataframe as bullet points.}
\usage{
print_achievements(text_data, bullet_style = c("-", "+"))
}
\description{
Print the achievements listed in a text dataframe as bullet points.
}
\seealso{
Other print: 
\code{\link{print_contact_info}()},
\code{\link{print_cover_body}()},
\code{\link{print_cover_plain}()},
\code{\link{print_latex_section}()},
\code{\link{print_resume_plain}()},
\code{\link{print_section}()},
\code{\link{print_txt_header}()},
\code{\link{print_txt_section}()}
}
\concept{print}
