% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{print_contact_info}
\alias{print_contact_info}
\title{Print contact info.}
\usage{
print_contact_info(
  contact_data,
  section = c("info", "links", "both"),
  sep = c(" | ", "\\n")
)
}
\description{
Print contact info.
}
\concept{print-utils}
