% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover.R
\name{print_cover_body}
\alias{print_cover_body}
\title{Print a cover letter from a text dataframe in a plain text or LaTeX style.}
\usage{
print_cover_body(
  text_data,
  position,
  company,
  name = NULL,
  style = c("latex", "txt"),
  use_bullets = FALSE,
  bullet_style = c("-", "+"),
  type = c("cover", "email")
)
}
\description{
Print a cover letter from a text dataframe in a plain text or LaTeX style.
}
\seealso{
Other print: 
\code{\link{print_achievements}()},
\code{\link{print_contact_info}()},
\code{\link{print_cover_plain}()},
\code{\link{print_latex_section}()},
\code{\link{print_resume_plain}()},
\code{\link{print_section}()},
\code{\link{print_txt_header}()},
\code{\link{print_txt_section}()}
}
\concept{print}
