% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover.R
\name{print_cover_plain}
\alias{print_cover_plain}
\title{Prepare and print each section of a plain text cover letter.}
\usage{
print_cover_plain(
  target = c("app", "base"),
  position,
  company,
  app_id = "latest",
  app_period = "latest",
  use_bullets = FALSE,
  bullet_style = c("-", "+"),
  type = c("cover", "email")
)
}
\description{
Prepare and print each section of a plain text cover letter.
}
\seealso{
Other print: 
\code{\link{print_achievements}()},
\code{\link{print_contact_info}()},
\code{\link{print_cover_body}()},
\code{\link{print_latex_section}()},
\code{\link{print_resume_plain}()},
\code{\link{print_section}()},
\code{\link{print_txt_header}()},
\code{\link{print_txt_section}()}
}
\concept{print}
