% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_resume.R
\name{print_latex_section}
\alias{print_latex_section}
\title{Print a section from a position dataframe to LaTeX.}
\usage{
print_latex_section(position_data, section_id, short_entries = FALSE)
}
\description{
Print a section from a position dataframe to LaTeX.
}
\seealso{
Other print: 
\code{\link{print_section}()},
\code{\link{print_txt_section}()}
}
\concept{print}
