% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plain.R
\name{print_resume_plain}
\alias{print_resume_plain}
\title{Prepare and print each section of a plain text resume.}
\usage{
print_resume_plain(
  target = c("app", "base"),
  app_id = "latest",
  app_period = "latest",
  use_abridged = FALSE
)
}
\description{
Prepare and print each section of a plain text resume.
}
\seealso{
Other print: 
\code{\link{print_achievements}()},
\code{\link{print_contact_info}()},
\code{\link{print_cover_body}()},
\code{\link{print_cover_plain}()},
\code{\link{print_latex_section}()},
\code{\link{print_section}()},
\code{\link{print_txt_header}()},
\code{\link{print_txt_section}()}
}
\concept{print}
