% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_plain.R
\name{print_txt_header}
\alias{print_txt_header}
\title{Print a formatted plain text resume header.}
\usage{
print_txt_header(header)
}
\description{
Print a formatted plain text resume header.
}
\examples{
print_txt_header("this is my header")

}
\seealso{
Other print: 
\code{\link{print_achievements}()},
\code{\link{print_contact_info}()},
\code{\link{print_cover_body}()},
\code{\link{print_cover_plain}()},
\code{\link{print_latex_section}()},
\code{\link{print_resume_plain}()},
\code{\link{print_section}()},
\code{\link{print_txt_section}()}
}
\concept{print}
