% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_plain.R
\name{print_txt_section}
\alias{print_txt_section}
\title{Print sections of a plain text resume.}
\usage{
print_txt_section(
  position_data,
  section_id = c("work", "education", "certifications", "projects", "writing",
    "publications", "volunteering", "additional_info"),
  target = c("app", "base"),
  use_abridged = FALSE
)
}
\description{
Print sections of a plain text resume.
}
\seealso{
Other print: 
\code{\link{print_achievements}()},
\code{\link{print_contact_info}()},
\code{\link{print_cover_body}()},
\code{\link{print_cover_plain}()},
\code{\link{print_latex_section}()},
\code{\link{print_resume_plain}()},
\code{\link{print_section}()},
\code{\link{print_txt_header}()}
}
\concept{print}
