% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_all}
\alias{render_all}
\title{Run each resume/CV rendering option in sequence.}
\usage{
render_all()
}
\description{
Run each resume/CV rendering option in sequence.
}
\seealso{
Other cli: 
\code{\link{build_app_directory}()},
\code{\link{get_app_info}()},
\code{\link{load_application_data}()},
\code{\link{render_cv_as_html}()}
}
\concept{cli}
