% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_app}
\alias{render_app}
\alias{render_base}
\alias{render_cv}
\alias{render_linkedin}
\title{Run each resume or CV rendering step in sequence.}
\usage{
render_app(
  app_id = "latest",
  app_period = "latest",
  docs = c("resume", "all", "cover"),
  use_bullets = TRUE,
  show = FALSE
)

render_base(report_counts = TRUE, cover = FALSE, show = FALSE)

render_cv(show = FALSE)

render_linkedin(report_counts = TRUE, show = FALSE)
}
\description{
\code{render_app} builds a pdf and plain text resume for a given application.

\code{render_base} builds an html CV, pdf and plain text resume from base data.

\code{render_cv} builds an html CV and converts it to pdf.

\code{render_linkedin} builds a short resume suitable for professional profiles.
}
\seealso{
Other cli: 
\code{\link{build_app_directory}()},
\code{\link{build_base_directory}()},
\code{\link{get_app_info}()},
\code{\link{open_app}()},
\code{\link{render_cover}()},
\code{\link{render_cv_as_html}()},
\code{\link{render_resume}()},
\code{\link{update_app_info}()}
}
\concept{cli}
