% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_cover}
\alias{render_cover}
\alias{render_cover_plain}
\title{Construct a cover letter based on spreadsheet data.}
\usage{
render_cover(
  target = c("app", "base"),
  app_id = "latest",
  app_period = "latest",
  use_bullets = TRUE,
  bullet_style = c("-", "+"),
  stylesheets = list("custom_cover.tex"),
  input_filename = "cover.Rmd",
  output_basename = "cover",
  app_dir = "applications",
  input_dir = "notebooks",
  data_dir = "input",
  output_dir = "output",
  show = FALSE
)

render_cover_plain(
  target = c("app", "base"),
  app_id = "latest",
  app_period = "latest",
  use_bullets = TRUE,
  bullet_style = c("-", "+"),
  type = c("cover", "email"),
  app_dir = "applications",
  data_dir = "input",
  output_dir = "output",
  show = FALSE
)
}
\description{
\code{render_cover} constructs a formal, LaTeX-style resume.

\code{render_cover_plain} constructs a plain text cover letter for simplicity.
}
\seealso{
Other cli: 
\code{\link{build_app_directory}()},
\code{\link{build_base_directory}()},
\code{\link{get_app_info}()},
\code{\link{open_app}()},
\code{\link{render_app}()},
\code{\link{render_cv_as_html}()},
\code{\link{render_resume}()},
\code{\link{update_app_info}()}
}
\concept{cli}
