% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_resume_plain}
\alias{render_resume_plain}
\title{Build an plain text resume based on spreadsheet data.}
\usage{
render_resume_plain(
  input_filename = "plain.Rmd",
  data_filename = "resume_data.xlsx",
  output_basename = "resume",
  input_dir = "notebooks",
  data_dir = "input",
  output_dir = "output"
)
}
\description{
Build an plain text resume based on spreadsheet data.
}
\seealso{
Other cli: 
\code{\link{build_app_directory}()},
\code{\link{get_app_info}()},
\code{\link{load_application_data}()},
\code{\link{render_all}()},
\code{\link{render_cv_as_html}()},
\code{\link{render_cv_as_pdf}()},
\code{\link{render_resume}()}
}
\concept{cli}
