% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{report_skill_metrics}
\alias{report_skill_metrics}
\alias{sort_skill_report}
\title{Display skill report info for at-a-glance viewing.}
\usage{
report_skill_metrics(skill_report_df)

sort_skill_report(skill_report_df)
}
\description{
\code{report_skill_metrics} computes basic coverage statistics.

\code{sort_skill_report} arranges rows in a convenient viewing order.
}
\seealso{
Other report: 
\code{\link{count_terms}()},
\code{\link{run_skill_count}()},
\code{\link{run_skill_report}()}
}
\concept{report}
