% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{set_project_paths}
\alias{set_project_paths}
\title{Establish the root directory when building a new project.}
\usage{
set_project_paths(relpath = NULL)
}
\arguments{
\item{relpath}{the desired project path relative to your working directory.}
}
\description{
Establish the root directory when building a new project.
}
\seealso{
Other build-dev: 
\code{\link{as_filename}()},
\code{\link{construct_app_metadata}()},
\code{\link{copy_files}()},
\code{\link{create_rprofile}()},
\code{\link{create_rproject}()},
\code{\link{get_latest_entry}()},
\code{\link{get_valid_opts}()},
\code{\link{in_tmp_env}()},
\code{\link{open_doc_and_wait}()},
\code{\link{open_with_application}()},
\code{\link{str_to_filename}()},
\code{\link{truncate_string}()},
\code{\link{validate_id}()},
\code{\link{validate_status}()},
\code{\link{write_app_metadata}()}
}
\concept{build-dev}
