% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{sort_skills}
\alias{sort_skills}
\title{Sort skills.}
\usage{
sort_skills(skill_data, target = c("app", "base"))
}
\description{
Sort skills.
}
\seealso{
Other prepare-dev: 
\code{\link{append_skills_to_bullets}()},
\code{\link{make_markdown_contacts}()},
\code{\link{omit_hidden_fields}()}
}
\concept{prepare-dev}
