% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{str_to_filename}
\alias{str_to_filename}
\title{Convert a string to a valid file name.}
\usage{
str_to_filename(string, sep = "-", lower = TRUE)
}
\description{
Convert a string to a valid file name.
}
\seealso{
Other build-dev: 
\code{\link{as_filename}()},
\code{\link{construct_app_metadata}()},
\code{\link{copy_files}()},
\code{\link{create_rprofile}()},
\code{\link{create_rproject}()},
\code{\link{get_latest_entry}()},
\code{\link{get_valid_opts}()},
\code{\link{in_tmp_env}()},
\code{\link{open_doc_and_wait}()},
\code{\link{open_with_application}()},
\code{\link{set_project_paths}()},
\code{\link{truncate_string}()},
\code{\link{validate_id}()},
\code{\link{validate_status}()},
\code{\link{write_app_metadata}()}
}
\concept{build-dev}
