% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{truncate_string}
\alias{truncate_string}
\title{Shorten long strings in a data frame for neater printing.}
\usage{
truncate_string(str, max_length = 20)
}
\description{
Shorten long strings in a data frame for neater printing.
}
\seealso{
Other build-dev: 
\code{\link{as_filename}()},
\code{\link{construct_app_metadata}()},
\code{\link{copy_files}()},
\code{\link{create_rprofile}()},
\code{\link{create_rproject}()},
\code{\link{get_latest_entry}()},
\code{\link{get_valid_opts}()},
\code{\link{in_tmp_env}()},
\code{\link{open_doc_and_wait}()},
\code{\link{open_with_application}()},
\code{\link{set_project_paths}()},
\code{\link{str_to_filename}()},
\code{\link{validate_id}()},
\code{\link{validate_status}()},
\code{\link{write_app_metadata}()}
}
\concept{build-dev}
