% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{update_app_info}
\alias{update_app_info}
\alias{update_datestamp}
\alias{apply_to}
\title{Update application metadata.}
\usage{
update_app_info(app_id = "latest", ..., log = load_log())

update_datestamp(app_id = "latest", log = load_log())

apply_to(app_id = "latest", log = load_log())
}
\arguments{
\item{...}{Key-value pairs of the fields to update and their new values.
If none are provided, the log is updated directly from the metadata file.}

\item{id}{The unique identifier associated with the application of interest.}
}
\description{
Update the log entry for a particular application as well its metadata file.

\code{update_app_info} updates the log and metadata from user-supplied data.

\code{update_datestamp} auto-updates the application datestamp upon rendering.

\code{apply_to} sets the \code{date_applied} datestamp and \code{status} to applied.
}
\seealso{
Other cli: 
\code{\link{build_app_directory}()},
\code{\link{build_base_directory}()},
\code{\link{get_app_info}()},
\code{\link{open_app}()},
\code{\link{render_app}()},
\code{\link{render_cover}()},
\code{\link{render_cv_as_html}()},
\code{\link{render_resume}()}
}
\concept{cli}
