% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{user.input}
\alias{user.input}
\title{Capture user input in both interactive and non-interactive sessions.}
\usage{
user.input(prompt)
}
\description{
Capture user input in both interactive and non-interactive sessions.
}
\keyword{internal}
