% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCV-docs.R, R/utils_.R
\name{utils_}
\alias{utils_}
\alias{not_empty}
\alias{messager}
\alias{stopper}
\alias{replace_extensions}
\title{Utils functions}
\usage{
not_empty(rc)

messager(..., v = TRUE, parallel = FALSE)

stopper(..., v = TRUE)

replace_extensions(files, replacement = "")
}
\arguments{
\item{rc}{Row data.}

\item{v}{Whether to print messages or not.}

\item{parallel}{Whether to enable message print when wrapped 
in parallelised functions.}

\item{files}{Character vector of file paths.}

\item{replacement}{a replacement for matched pattern in \code{sub} and
    \code{gsub}.  Coerced to character if possible.  For
    \code{fixed = FALSE} this can include backreferences \code{"\1"} to
    \code{"\9"} to parenthesized subexpressions of \code{pattern}.  For
    \code{perl = TRUE} only, it can also contain \code{"\U"} or
    \code{"\L"} to convert the rest of the replacement to upper or
    lower case and \code{"\E"} to end case conversion.  If a
    character vector of length 2 or more is supplied, the first element
    is used with a warning.  If \code{NA}, all elements in the result
    corresponding to matches will be set to \code{NA}.
  }
}
\value{
Null.

boolean

Null

Null

Character vector
}
\description{
Internal utility functions.
}
\section{Functions}{
\itemize{
\item \code{not_empty()}: Check if a field is not empty.

\item \code{messager()}: Conditionally print messages.
Allows developers to easily control verbosity of functions, 
and meet Bioconductor requirements that dictate the message 
must first be stored to a variable before passing to \link[base]{message}.

\item \code{stopper()}: Conditionally print stop messages.
Allows developers to easily control verbosity of functions, 
 and meet Bioconductor requirements that dictate the stop message 
 must first be stored to a variable before passing to \link[base]{stop}.

\item \code{replace_extensions()}: Replace the extension name for a set of files.

}}
\concept{utils_}
\keyword{internal}
