% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{write_log_entry}
\alias{write_log_entry}
\title{Record the current application in the log for the present period.}
\usage{
write_log_entry(app_df = construct_app_metadata())
}
\description{
Record the current application in the log for the present period.
}
\seealso{
Other cli-utils: 
\code{\link{construct_app_filepath}()},
\code{\link{construct_app_metadata}()},
\code{\link{load_job_info}()},
\code{\link{load_log}()},
\code{\link{str_to_filename}()},
\code{\link{validate_id}()},
\code{\link{write_app_metadata}()}
}
\concept{cli-utils}
