% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_alerts.R
\name{alert_opening}
\alias{alert_opening}
\title{Issue an alert that a file or folder is opening from the provided path.}
\usage{
alert_opening(path, base_dir = ".", action = "")
}
\arguments{
\item{path}{an absolute path to the file or folder in question.}
}
\description{
Issue an alert that a file or folder is opening from the provided path.
}
\seealso{
Other cli-dev: 
\code{\link{alert_file_created}()},
\code{\link{alert_writing_to}()},
\code{\link{format_action}()},
\code{\link{warn_file_exists}()},
\code{\link{warn_file_missing}()}
}
\concept{cli-dev}
