% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{append_skills_to_bullets}
\alias{append_skills_to_bullets}
\title{Append skills to a description field with matching index.}
\usage{
append_skills_to_bullets(data, ix, use_abridged = FALSE, omit_prefix = "/")
}
\description{
Append skills to a description field with matching index.
}
\seealso{
Other prepare-dev: 
\code{\link{make_markdown_contacts}()},
\code{\link{omit_hidden_fields}()},
\code{\link{sort_skills}()}
}
\concept{prepare-dev}
