% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{autocv_resources}
\alias{autocv_resources}
\title{Find the location of template data files on your system.}
\usage{
autocv_resources(path = NULL, inst_dir = "extdata")
}
\description{
Find the location of template data files on your system.
}
\examples{
# Obtain a complete list of top-level resources included with autocv:
autocv_resources()

# Obtain a complete list of resources included in the css/ directory:
autocv_resources("css")

# Obtain a complete path to the custom_resume.css resource:
autocv_resources("css/custom_resume.css")

# Obtain a complete list of template files included with autocv:
autocv_resources(inst_dir = "templates")

}
\seealso{
Other data: 
\code{\link{get_path_to}()}
}
\concept{data}
