% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_app_directory}
\alias{build_app_directory}
\title{Create the file tree and data files for the present app.}
\usage{
build_app_directory(app_df = construct_app_metadata(), base_app_id = NA)
}
\description{
Create the file tree and data files for the present app.
}
\seealso{
Other cli: 
\code{\link{get_app_info}()},
\code{\link{load_application_data}()},
\code{\link{render_all}()},
\code{\link{render_cv_as_html}()},
\code{\link{render_cv_as_pdf}()},
\code{\link{render_resume}()},
\code{\link{render_resume_plain}()}
}
\concept{cli}
