% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{cli_build}
\alias{cli_build}
\title{CLI handler for 'build'.}
\usage{
cli_build(args)
}
\description{
CLI handler for 'build'.
}
\seealso{
Other cli-handlers: 
\code{\link{cli_edit}()},
\code{\link{cli_render}()}
}
\concept{cli-handlers}
