% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{construct_app_metadata}
\alias{construct_app_metadata}
\title{Create the metadata for a new application.}
\usage{
construct_app_metadata(
  metadata_filename = "job_metadata.yml",
  metadata_dir = "input"
)
}
\description{
This information is also used to construct the application log entry.
}
\seealso{
Other cli-utils: 
\code{\link{construct_app_filepath}()},
\code{\link{load_job_info}()},
\code{\link{load_log}()},
\code{\link{str_to_filename}()},
\code{\link{validate_id}()},
\code{\link{write_app_metadata}()},
\code{\link{write_log_entry}()}
}
\concept{cli-utils}
