% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{count_terms}
\alias{count_terms}
\title{Count keyword occurrences in a file based on a supplied list.}
\usage{
count_terms(
  terms,
  counts = NULL,
  doc,
  orderby = c("counts", "doc", "source"),
  filterby = c("both", "count", "matches")
)
}
\description{
Count keyword occurrences in a file based on a supplied list.
}
\examples{
# Load example data
data("example_posting", package = "autocv")
terms <- readLines(autocv_resources("resources/skill_list.txt"))

# Generate some random counts
counts <- sample(0:10, length(prep_term_list(terms)), replace = TRUE)

posting_counts <- count_terms(
  terms = terms,
  counts = counts,
  doc = example_posting,
  orderby = "doc",
  filterby = "both"
)
posting_counts

}
\seealso{
Other report: 
\code{\link{report_skill_metrics}()},
\code{\link{run_skill_count}()},
\code{\link{run_skill_report}()}
}
\concept{report}
