% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_skill_data}
\alias{example_skill_data}
\title{Example skill dataset}
\format{
\subsection{\code{example_skill_data}}{

A data frame with x rows and y columns:
\describe{
\item{field}{Field description}
\item{field 1, field 2}{Field descriptions}
\item{year}{Year}
}
}
}
\source{
\url{https://source}
}
\usage{
example_skill_data
}
\description{
A subset of data from the ...
Report ...
}
\keyword{internal}
