% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_text_data}
\alias{example_text_data}
\title{Sample text data}
\format{
A data frame with 32 rows and 7 columns:
\describe{
\item{loc}{Name of the text block (character).}

\item{include}{Flag indicating whether the item should be included
in your tailored resume, either "x" or NA (character).}

\item{text}{The actual text to display for the item (character).}

\item{order}{Ordinal position in which to display the item
in your application documents (positive integer).}

\item{word_limit}{Limit to impose on the number of words
for the text block (positive integer).}

\item{word_count}{Computed number of words
for the text block; auto-populated (positive integer).}
}
}
\source{
Generated for demonstration purposes.
}
\usage{
example_text_data
}
\description{
A synthetic dataset containing the textual building blocks of
job application components such as a resume bio or a cover letter.
For more details, refer to the corresponding
\href{https://github.com/bainmatt/autocv/blob/main/inst/templates/template_cover_data.xlsx}{template spreadsheet}
in the repository.
}
\examples{
data(example_text_data)
head(example_text_data)

}
\keyword{internal}
