% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{get_app_info}
\alias{get_app_info}
\alias{get_path_to_app}
\alias{open_app}
\title{Access and open application files.}
\usage{
get_app_info(
  id = "all",
  field = "all",
  company = NA,
  position = NA,
  app_dir = "applications",
  app_period = "latest"
)

get_path_to_app(
  id,
  document,
  company = NA,
  position = NA,
  app_dir = "applications",
  app_period = "latest"
)

open_app(
  id = "latest",
  document = "resume",
  company = NA,
  position = NA,
  app_dir = "applications",
  app_period = "latest"
)
}
\description{
\code{get_app_info} retrieves requested info for a given application from its log.

\code{get_path_to_app} returns an abs path to a document from a given application.

\code{open_app} opens the requested document for a given application.
}
\seealso{
Other cli: 
\code{\link{build_app_directory}()},
\code{\link{load_application_data}()},
\code{\link{render_all}()},
\code{\link{render_cv_as_html}()},
\code{\link{render_cv_as_pdf}()},
\code{\link{render_resume}()},
\code{\link{render_resume_plain}()}
}
\concept{cli}
