% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{get_latest_entry}
\alias{get_latest_entry}
\title{Obtain the latest entry in an application log.}
\usage{
get_latest_entry(log = load_log())
}
\description{
Obtain the latest entry in an application log.
}
\seealso{
Other build-dev: 
\code{\link{as_filename}()},
\code{\link{construct_app_metadata}()},
\code{\link{copy_files}()},
\code{\link{create_rprofile}()},
\code{\link{create_rproject}()},
\code{\link{get_valid_opts}()},
\code{\link{in_tmp_env}()},
\code{\link{open_doc_and_wait}()},
\code{\link{set_project_paths}()},
\code{\link{str_to_filename}()},
\code{\link{truncate_string}()},
\code{\link{validate_id}()},
\code{\link{validate_status}()},
\code{\link{write_app_metadata}()}
}
\concept{build-dev}
