% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{load_application_data}
\alias{load_application_data}
\title{Load application data (resume or cover letter).}
\usage{
load_application_data(
  data_dir = "input",
  filename = "resume_data.xlsx",
  sheet = c("entries", "skills", "contact_info", "text_blocks"),
  skip = 1
)
}
\description{
Load application data (resume or cover letter).
}
\seealso{
Other cli: 
\code{\link{build_app_directory}()},
\code{\link{get_app_info}()},
\code{\link{render_all}()},
\code{\link{render_cv_as_html}()},
\code{\link{render_cv_as_pdf}()},
\code{\link{render_resume}()},
\code{\link{render_resume_plain}()}
}
\concept{cli}
