% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{load_application_data}
\alias{load_application_data}
\title{Load application data (resume or cover letter).}
\usage{
load_application_data(
  target = c("app", "base"),
  filename = c("resume_data.xlsx", "cover_data.xlsx"),
  sheet = c("entries", "skills", "contact_info", "text_blocks"),
  data_dir = "input",
  skip = 1,
  app_id = "latest",
  app_period = "latest",
  app_dir = "applications"
)
}
\description{
Load application data (resume or cover letter).
}
\seealso{
Other cli: 
\code{\link{build_app_directory}()},
\code{\link{get_app_info}()},
\code{\link{render_all}()},
\code{\link{render_cv_as_html}()}
}
\concept{cli}
