% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{load_job_info}
\alias{load_job_info}
\title{Read user-defined metadata for the present job from a config file.}
\usage{
load_job_info(field = "all", filename = "job_metadata.yml", dir = "input")
}
\value{
A nested named list of strings with user-specified job details
}
\description{
Read user-defined metadata for the present job from a config file.
}
\seealso{
Other cli-utils: 
\code{\link{construct_app_filepath}()},
\code{\link{construct_app_metadata}()},
\code{\link{load_log}()},
\code{\link{str_to_filename}()},
\code{\link{validate_id}()},
\code{\link{write_app_metadata}()},
\code{\link{write_log_entry}()}
}
\concept{cli-utils}
