% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{load_log}
\alias{load_log}
\title{Load the log for a given application period.}
\usage{
load_log(app_dir = "applications", app_period = "latest")
}
\description{
Load the log for a given application period.
}
\seealso{
Other cli-utils: 
\code{\link{construct_app_filepath}()},
\code{\link{construct_app_metadata}()},
\code{\link{load_job_info}()},
\code{\link{str_to_filename}()},
\code{\link{validate_id}()},
\code{\link{write_app_metadata}()},
\code{\link{write_log_entry}()}
}
\concept{cli-utils}
