% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{omit_hidden_fields}
\alias{omit_hidden_fields}
\title{Omit spreadsheet entries beginning with a preset prefix.}
\usage{
omit_hidden_fields(data, style = "markdown", prefix = "/")
}
\description{
Omit spreadsheet entries beginning with a preset prefix.
}
\seealso{
Other prepare-utils: 
\code{\link{append_skills_to_bullets}()},
\code{\link{make_markdown_contacts}()},
\code{\link{sort_skills}()}
}
\concept{prepare-utils}
