% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCV-docs.R, R/parse_.R
\name{parse_}
\alias{parse_}
\alias{parse_location}
\alias{parse_daterange}
\alias{parse_bullets}
\alias{parse_news}
\alias{parse_education}
\alias{parse_publications}
\alias{parse_talks}
\alias{parse_experience}
\alias{parse_tools}
\alias{parse_profile}
\alias{parse_affiliations}
\alias{parse_grants_totals}
\alias{parse_grants}
\alias{parse_skills}
\title{Parse functions}
\usage{
parse_location(r, add_link = TRUE)

parse_daterange(r)

parse_bullets(r, concise = FALSE, check_icons = TRUE)

parse_news(r, title = "News", icon = "newspaper")

parse_education(
  wd = "./",
  file = file.path(wd, "cv_data", "education.csv"),
  concise = FALSE
)

parse_publications(
  file = file.path("cv_data", "publications.csv"),
  name = "Leonardo da Vinci",
  types = NULL
)

parse_talks(file = file.path("cv_data", "talks.csv"), types = NULL)

parse_experience(
  file = file.path("cv_data", "experience.csv"),
  types = NULL,
  concise = FALSE
)

parse_tools(
  wd = "./",
  file = file.path(wd, "cv_data", "tools.csv"),
  types = NULL,
  add_index = TRUE,
  add_logos = TRUE
)

parse_profile(
  wd = "./",
  file = file.path(wd, "cv_data", "profile.csv"),
  types = NULL,
  concise = FALSE,
  sep = "\\n\\n",
  collapse = "<br>",
  div = "h4",
  img_width = "100px",
  icn_width = "12px",
  prefix = NULL
)

parse_affiliations(
  wd = "./",
  types = "affiliation",
  sep = "\\n\\n",
  collapse = "\\n\\n",
  div = NULL,
  img_width = "150px",
  prefix = "### "
)

parse_grants_totals(dt)

parse_grants(
  file = file.path("cv_data", "grants.csv"),
  types = NULL,
  add_totals = FALSE
)

parse_skills(file = file.path("cv_data", "skills.csv"), types = NULL)
}
\arguments{
\item{r}{One row from a \link[data.table]{data.table}.}

\item{add_link}{Add link.}

\item{concise}{Use the concise layout.}

\item{check_icons}{Check icons.}

\item{title}{Title.}

\item{icon}{Icon.}

\item{wd}{Working directory where the "cv_data" and "img" subfolders
are stored.}

\item{file}{Data file path.}

\item{name}{Your name to put on the CV.}

\item{types}{Subset the `Type` column.}

\item{add_index}{Add index.}

\item{add_logos}{Add logos.}

\item{sep}{a character string to separate the terms.  Not
    \code{\link[base]{NA_character_}}.}

\item{collapse}{an optional character string to separate the results.  Not
    \code{\link[base]{NA_character_}}.}

\item{div}{HTML div class.}

\item{img_width}{Image width.}

\item{icn_width}{Icon width.}

\item{prefix}{Prefix.}

\item{dt}{Grants \link[data.table]{data.table}.}

\item{add_totals}{Add totals.}
}
\value{
Formatted data.
}
\description{
Functions to parse rows from your CV data.
}
\section{Functions}{
\itemize{
\item \code{parse_location()}: parse_

\item \code{parse_daterange()}: parse_

\item \code{parse_bullets()}: parse_

\item \code{parse_news()}: parse_

\item \code{parse_education()}: parse_

\item \code{parse_publications()}: parse_

\item \code{parse_talks()}: parse_

\item \code{parse_experience()}: parse_

\item \code{parse_tools()}: parse_

\item \code{parse_profile()}: parse_

\item \code{parse_affiliations()}: parse_

\item \code{parse_grants_totals()}: parse_

\item \code{parse_grants()}: parse_

\item \code{parse_skills()}: parse_

}}
\concept{parse_}
