% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_links}
\alias{prepare_links}
\title{Prepare properly formatted links with custom text.}
\usage{
prepare_links(
  data,
  style = c("markdown", "latex", "txt"),
  macro = c("myhref", "href", NA)
)
}
\arguments{
\item{style}{style of link, either "markdown", "latex", or "txt" (plain)}

\item{macro}{macro to use for "latex" references}
}
\description{
Prepare properly formatted links with custom text.
}
\seealso{
Other prepare: 
\code{\link{prepare_bio}()},
\code{\link{prepare_description_bullets}()},
\code{\link{prepare_timeline}()}
}
\concept{prepare}
