% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{print_contact_info}
\alias{print_contact_info}
\title{Print contact info.}
\usage{
print_contact_info(
  contact_data,
  section = c("info", "links", "both", "signoff"),
  sep = c(" | ", "\\n"),
  anonymize = FALSE
)
}
\description{
Print contact info.
}
\seealso{
Other print: 
\code{\link{print_achievements}()},
\code{\link{print_cover_body}()},
\code{\link{print_cover_plain}()},
\code{\link{print_latex_section}()},
\code{\link{print_resume_plain}()},
\code{\link{print_section}()},
\code{\link{print_txt_header}()},
\code{\link{print_txt_section}()}
}
\concept{print}
