% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_cv.R
\name{print_section}
\alias{print_section}
\title{Print a section from a position dataframe to markdown.}
\usage{
print_section(position_data, section_id)
}
\description{
Print a section from a position dataframe to markdown.
}
\seealso{
Other print: 
\code{\link{print_latex_section}()},
\code{\link{print_txt_section}()}
}
\concept{print}
