% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_plain.R
\name{print_txt_section}
\alias{print_txt_section}
\title{Print sections of a plain text resume.}
\usage{
print_txt_section(position_data)
}
\description{
Print sections of a plain text resume.
}
\seealso{
Other print: 
\code{\link{print_latex_section}()},
\code{\link{print_section}()}
}
\concept{print}
