% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_cv_as_html}
\alias{render_cv_as_html}
\alias{render_cv_as_pdf}
\title{Construct a curriculum vitae based on spreadsheet data.}
\usage{
render_cv_as_html(
  input_filename = "cv.Rmd",
  output_basename = "cv",
  input_dir = "notebooks",
  data_dir = "input",
  output_dir = "output",
  stylesheets = list("custom_resume.css", "styles_html.css"),
  show = FALSE
)

render_cv_as_pdf(
  input_filename = "cv.Rmd",
  output_basename = "cv",
  input_dir = "notebooks",
  data_dir = "input",
  output_dir = "output",
  stylesheets = list("custom_resume.css", "styles_pdf.css")
)
}
\description{
\code{render_cv_as_html} constructs an informal HTML-format resume.

\code{render_cv_as_pdf} converts an HTML-format resume to PDF.
}
\seealso{
Other cli: 
\code{\link{build_app_directory}()},
\code{\link{build_base_directory}()},
\code{\link{get_app_info}()},
\code{\link{open_app}()},
\code{\link{render_app}()},
\code{\link{render_cover}()},
\code{\link{render_resume}()},
\code{\link{update_app_info}()}
}
\concept{cli}
