% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_cv_as_html}
\alias{render_cv_as_html}
\title{Build an informal HTML-format resume based on spreadsheet data.}
\usage{
render_cv_as_html(
  input_filename = "cv.Rmd",
  data_filename = "resume_data.xlsx",
  output_basename = "cv",
  input_dir = "notebooks",
  data_dir = "input",
  output_dir = "output",
  stylesheets = list("custom_resume.css", "styles_html.css")
)
}
\description{
Build an informal HTML-format resume based on spreadsheet data.
}
\seealso{
Other cli: 
\code{\link{build_app_directory}()},
\code{\link{get_app_info}()},
\code{\link{load_application_data}()},
\code{\link{render_all}()},
\code{\link{render_cv_as_pdf}()},
\code{\link{render_resume}()},
\code{\link{render_resume_plain}()}
}
\concept{cli}
