% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_cv_as_html}
\alias{render_cv_as_html}
\alias{render_cv_as_pdf}
\alias{render_resume}
\alias{render_resume_plain}
\title{Construct a resume based on spreadsheet data.}
\usage{
render_cv_as_html(
  input_filename = "cv.Rmd",
  data_filename = "resume_data.xlsx",
  output_basename = "cv",
  input_dir = "notebooks",
  data_dir = "input",
  output_dir = "output",
  stylesheets = list("custom_resume.css", "styles_html.css")
)

render_cv_as_pdf(
  input_filename = "cv.Rmd",
  data_filename = "resume_data.xlsx",
  output_basename = "cv",
  input_dir = "notebooks",
  data_dir = "input",
  output_dir = "output",
  stylesheets = list("custom_resume.css", "styles_pdf.css")
)

render_resume(
  target = c("app", "base"),
  input_filename = "resume.Rmd",
  data_filename = "resume_data.xlsx",
  output_basename = "resume",
  input_dir = "notebooks",
  data_dir = "input",
  output_dir = "output",
  stylesheets = list("custom_resume.tex"),
  app_id = "latest",
  app_period = "latest",
  app_dir = "applications"
)

render_resume_plain(
  target = c("app", "base"),
  input_filename = "plain.Rmd",
  data_filename = "resume_data.xlsx",
  output_basename = "resume",
  input_dir = "notebooks",
  data_dir = "input",
  output_dir = "output"
)
}
\description{
\code{render_cv_as_html} constructs an informal HTML-format resume.

\code{render_cv_as_pdf} converts an HTML-format resume to PDF.

\code{render_resume} constructs a formal, ATS-compatible, LaTeX-style resume.

\code{render_resume_plain} constructs a plain text resume for analysis.
}
\seealso{
Other cli: 
\code{\link{build_app_directory}()},
\code{\link{get_app_info}()},
\code{\link{load_application_data}()},
\code{\link{render_all}()}
}
\concept{cli}
