% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_resume}
\alias{render_resume}
\alias{render_resume_plain}
\title{Construct a resume based on spreadsheet data.}
\usage{
render_resume(
  target = c("app", "base"),
  app_id = "latest",
  app_period = "latest",
  stylesheets = list("custom_resume.tex"),
  input_filename = "resume.Rmd",
  output_basename = "resume",
  app_dir = "applications",
  input_dir = "notebooks",
  data_dir = "input",
  output_dir = "output",
  use_abridged = FALSE
)

render_resume_plain(
  target = c("app", "base"),
  app_id = "latest",
  app_period = "latest",
  output_basename = "resume",
  app_dir = "applications",
  data_dir = "input",
  output_dir = "output",
  use_abridged = FALSE
)
}
\description{
\code{render_resume} constructs a formal, ATS-compatible, LaTeX-style resume.
}
\details{
\code{render_resume_plain} constructs a plain text resume for analysis.
}
\examples{
# TODO: just use templates & base

}
\seealso{
Other cli: 
\code{\link{build_app_directory}()},
\code{\link{build_base_directory}()},
\code{\link{get_app_info}()},
\code{\link{open_app}()},
\code{\link{render_app}()},
\code{\link{render_cover}()},
\code{\link{render_cv_as_html}()},
\code{\link{update_app_info}()}
}
\concept{cli}
