% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_skills_report.R
\name{run_skill_count}
\alias{run_skill_count}
\alias{check_skills}
\title{Obtain keyword counts for a given doc/term list and generate reports.}
\usage{
run_skill_count(
  app_id,
  app_period = "latest",
  doc = c("posting", "resume"),
  term_list_filename = c("skill_list.txt", "keyword_list.txt"),
  term_list_dir = "resources"
)

check_skills(
  app_id,
  app_period = "latest",
  doc = list("posting", "resume"),
  term_list_filename = c("skill_list.txt", "keyword_list.txt"),
  term_list_dir = "resources"
)
}
\description{
\code{run_skill_count} returns keyword counts for a given doc/term list.

\code{check_skills} obtains skill counts for a posting and creates a resume report
}
\seealso{
Other report: 
\code{\link{count_terms}()},
\code{\link{run_skill_report}()}
}
\concept{report}
