% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{run_skill_report}
\alias{run_skill_report}
\title{Generate a skill report for a given job application.}
\usage{
run_skill_report(
  app_id = "latest",
  log = load_log(),
  orderby = c("source", "counts", "doc")
)
}
\description{
Generate a skill report for a given job application.
}
\seealso{
Other report: 
\code{\link{count_terms}()},
\code{\link{report_skill_metrics}()},
\code{\link{run_skill_count}()}
}
\concept{report}
