% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{str_to_filename}
\alias{str_to_filename}
\title{Convert a string to a valid file name.}
\usage{
str_to_filename(string, separator = "-", lower = TRUE)
}
\description{
Convert a string to a valid file name.
}
\seealso{
Other cli-utils: 
\code{\link{construct_app_filepath}()},
\code{\link{construct_app_metadata}()},
\code{\link{load_job_info}()},
\code{\link{load_log}()},
\code{\link{validate_id}()},
\code{\link{write_app_metadata}()},
\code{\link{write_log_entry}()}
}
\concept{cli-utils}
