% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCV-templates.R, R/utils_.R
\name{utils_}
\alias{utils_}
\alias{not_empty}
\alias{messager}
\alias{stopper}
\title{Utils functions}
\usage{
not_empty(rc)

messager(..., v = TRUE, parallel = FALSE)

stopper(..., v = TRUE)
}
\arguments{
\item{rc}{Row data.}

\item{v}{Whether to print messages or not.}

\item{parallel}{Whether to enable message print when wrapped 
in parallelised functions.}
}
\value{
Null.

boolean

Null

Null
}
\description{
Internal utility functions.
}
\section{Functions}{
\itemize{
\item \code{not_empty()}: Check if a field is not empty.

\item \code{messager()}: Conditionally print messages.
Allows developers to easily control verbosity of functions, 
and meet Bioconductor requirements that dictate the message 
must first be stored to a variable before passing to \link[base]{message}.

\item \code{stopper()}: Conditionally print stop messages.
Allows developers to easily control verbosity of functions, 
 and meet Bioconductor requirements that dictate the stop message 
 must first be stored to a variable before passing to \link[base]{stop}.

}}
\concept{utils_}
\keyword{internal}
