% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{validate_id}
\alias{validate_id}
\title{Verify that matching job identifiers belong to the same job.}
\usage{
validate_id(log_data, app_df)
}
\description{
Verify that matching job identifiers belong to the same job.
}
\seealso{
Other cli-utils: 
\code{\link{construct_app_filepath}()},
\code{\link{construct_app_metadata}()},
\code{\link{load_job_info}()},
\code{\link{load_log}()},
\code{\link{str_to_filename}()},
\code{\link{write_app_metadata}()},
\code{\link{write_log_entry}()}
}
\concept{cli-utils}
