% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{validate_status}
\alias{validate_status}
\title{Verify that the supplied status is valid.}
\usage{
validate_status(
  status = c("all", "ipr", "applied", "interviewed", "rejected",
    "interviewed_then_rejected", "closed"),
  several.ok = FALSE
)
}
\description{
Verify that the supplied status is valid.
}
\seealso{
Other build-dev: 
\code{\link{as_filename}()},
\code{\link{construct_app_metadata}()},
\code{\link{copy_files}()},
\code{\link{create_rprofile}()},
\code{\link{create_rproject}()},
\code{\link{get_latest_entry}()},
\code{\link{get_valid_opts}()},
\code{\link{in_tmp_env}()},
\code{\link{open_doc_and_wait}()},
\code{\link{set_project_paths}()},
\code{\link{str_to_filename}()},
\code{\link{truncate_string}()},
\code{\link{validate_id}()},
\code{\link{write_app_metadata}()}
}
\concept{build-dev}
