% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_alerts.R
\name{warn_file_exists}
\alias{warn_file_exists}
\alias{warn_folder_exists}
\title{Issue a warning that a file or folder already exists at the provided path.}
\usage{
warn_file_exists(file, base_dir = ".", action = "skipping")

warn_folder_exists(folder, base_dir = ".", action = "skipping")
}
\description{
\code{warn_file_exists} issues a warning that a file already exists.

\code{warn_folder_exists} issues a warning that a folder already exists.
}
\seealso{
Other cli-dev: 
\code{\link{alert_file_created}()},
\code{\link{alert_opening}()},
\code{\link{alert_writing_to}()},
\code{\link{format_action}()},
\code{\link{warn_file_missing}()}
}
\concept{cli-dev}
