% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_alerts.R
\name{warn_file_missing}
\alias{warn_file_missing}
\alias{warn_folder_missing}
\title{Issue a warning that a file or folder does not exist at the provided path.}
\usage{
warn_file_missing(file, base_dir = ".", action = "aborting")

warn_folder_missing(folder, base_dir = ".", action = "aborting")
}
\description{
\code{warn_file_missing} issues a warning that a file is missing.

\code{warn_folder_missing} issues a warning that a folder is missing.
}
\seealso{
Other cli-dev: 
\code{\link{alert_file_created}()},
\code{\link{alert_opening}()},
\code{\link{alert_writing_to}()},
\code{\link{format_action}()},
\code{\link{warn_file_exists}()}
}
\concept{cli-dev}
