% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{write_app_metadata}
\alias{write_app_metadata}
\title{Save the metadata for a new application.}
\usage{
write_app_metadata(app_df = construct_app_metadata())
}
\description{
Save the metadata for a new application.
}
\seealso{
Other cli-utils: 
\code{\link{construct_app_filepath}()},
\code{\link{construct_app_metadata}()},
\code{\link{load_job_info}()},
\code{\link{load_log}()},
\code{\link{str_to_filename}()},
\code{\link{validate_id}()},
\code{\link{write_log_entry}()}
}
\concept{cli-utils}
